refactor(loon-core): move namespace catalog helpers out of services
