+++ crates/loon-core/src/context.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct MutationContext {
+    pub writer_id: String,
+    pub writer_version: String,
+    pub now_ms: u64,
+    pub lease_duration_ms: u64,
+}
