+++ crates/loon-core/src/lib.rs
@@ -4,6 +4,8 @@ mod basis;
 mod checkpoint;
 pub mod commit;
 pub mod content;
+mod context;
+mod error;
 mod genesis;
 pub mod invariants;
 mod lease;
@@ -18,6 +20,8 @@ pub use basis::{load_verified_namespace_basis, BasisLoadError, VerifiedNamespace
 pub use checkpoint::{
     advance_retention_floor, create_checkpoint, CheckpointLoadError, CheckpointLoadErrorKind,
 };
+pub use context::MutationContext;
+pub use error::{CoreError, CoreErrorKind};
 pub use lease::{acquire_or_renew_namespace_lease, LeaseAcquireError};
 pub use protocol::{
     begin_upload, commit_operations, complete_upload, list_changes_after, upload_content,
@@ -25,6 +29,6 @@ pub use protocol::{
 pub use services::{
     bootstrap_namespace, copy_file_path, delete_path, delete_path_non_recursive, fork_namespace,
     list_namespaces, list_path, move_path, put_file_bytes, put_file_content_ref, read_file_bytes,
-    resolve_path, store_bytes_as_content, write_file_bytes, BootstrapNamespaceError, CoreError,
-    CoreErrorKind, MutationContext, PutFileBehavior, StoredContent,
+    resolve_path, store_bytes_as_content, write_file_bytes, BootstrapNamespaceError,
+    PutFileBehavior, StoredContent,
 };
